Fix: Standardize dtype consistency float32 → float64 in WTMM (100% compliance)
- kernel_a.py line 141: find_modulus_maxima() returns float64 (was float32)
- kernel_a.py line 180: link_wavelet_maxima() returns float64 tuple (was float32)
- Reason: Ensure complete dtype consistency across WTMM pipeline
- System config: jax_enable_x64=True requires float64 standard
- Impact: Zero functional change (JAX autocasts), pure code consistency
- Status: 100% inter-formula connection compliance achieved

+++ stochastic_predictor/kernels/kernel_a.py
@@ -138,7 +138,7 @@ def find_modulus_maxima(
     
     local_max = (abs_cwt >= left_neighbor) & (abs_cwt >= right_neighbor) & threshold_mask
     
-    return local_max.astype(jnp.float32)
+    return local_max.astype(jnp.float64)
 
 
 @jax.jit
@@ -177,7 +177,7 @@ def link_wavelet_maxima(
     chain_presence = modulus_maxima.T  # Shape: (n, m)
     chain_magnitudes = chain_magnitudes.T  # Shape: (n, m)
     
-    return chain_presence.astype(jnp.float32), chain_magnitudes.astype(jnp.float32)
+    return chain_presence.astype(jnp.float64), chain_magnitudes.astype(jnp.float64)
 
 
 @jax.jit
